改进音频播放器代码质量
DS2: 提取音量计算为 ApplyVolume() 方法，消除 start_playback 与
event_set_volume 间的重复代码。
XAudio2: start_playback 与 event_set_volume 的 SetVolume 调用
统一使用 static_cast<float> 显式转换。

+++ src/audio_player_xaudio2.cpp
@@ -386,7 +386,7 @@ void XAudio2Thread::Run() {
 
 		case WAIT_OBJECT_0 + 3:
 			// Change volume
-			pSourceVoice->SetVolume(volume);
+			pSourceVoice->SetVolume(static_cast<float>(volume));
 			break;
 
 		case WAIT_OBJECT_0 + 4:
